# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A composer for Research Objects",
    "version": "0.1.0",
    "title": "RO Composer",
    "contact": {
      "email": "finn.bacall@manchester.ac.uk"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "cspool55.cs.man.ac.uk:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "profile",
      "description": "A profile for describing Research Objects"
    },
    {
      "name": "research_object",
      "description": "Research Objects"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/profiles": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "List Profiles",
        "description": "Show a list of available Research Object Profiles",
        "operationId": "listProfiles",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/hal+json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ProfileList"
            }
          }
        }
      }
    },
    "/profiles/{name}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "description": "Get a Research Object Profile",
        "operationId": "getProfile",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "The unique identifier of the profile",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/hal+json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "404": {
            "description": "Profile not found"
          }
        }
      }
    },
    "/profiles/{name}/research_objects": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "List Research Objects for Profile",
        "description": "Show a list of Research Objects that conform to the given profile",
        "operationId": "listResearchObjectsForProfile",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "The unique identifier of the profile",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/hal+json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ResearchObjectSummaryList"
            }
          },
          "404": {
            "description": "Profile not found"
          }
        }
      },
      "post": {
        "tags": [
          "research_object"
        ],
        "summary": "Create Research Object",
        "description": "Create a new Research Object that conforms to the given profile",
        "operationId": "createResearchObject",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The content of the Research Object (optional)",
            "schema": {
              "type": "object"
            }
          },
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "The unique identifier of the profile",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/hal+json"
        ],
        "responses": {
          "201": {
            "description": "successfully created",
            "schema": {
              "$ref": "#/definitions/ResearchObject"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Profile not found"
          }
        }
      }
    },
    "/research_objects": {
      "get": {
        "tags": [
          "research_object"
        ],
        "summary": "List Research Objects",
        "description": "Show a list of all Research Objects",
        "operationId": "listResearchObjects",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/hal+json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ResearchObjectSummaryList"
            }
          }
        }
      }
    },
    "/research_objects/{id}": {
      "get": {
        "tags": [
          "research_object"
        ],
        "summary": "Get Research Object",
        "description": "Get a Research Objects",
        "operationId": "getResearchObject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/hal+json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ResearchObject"
            }
          },
          "404": {
            "description": "Research object not found"
          }
        }
      }
    },
    "/research_objects/{id}/content/{field}": {
      "get": {
        "tags": [
          "research_object"
        ],
        "summary": "Get Research Object field",
        "description": "Retrieve a single field of the Research Object",
        "operationId": "getResearchObjectField",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          },
          {
            "in": "path",
            "name": "field",
            "type": "string",
            "description": "The name of the field to get, see profile for a list of field names",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "The given field is not present on the Research Object"
          },
          "404": {
            "description": "Research object not found"
          }
        }
      },
      "put": {
        "tags": [
          "research_object"
        ],
        "summary": "Set Research Object field",
        "description": "Set a single field on the Research Object",
        "operationId": "setResearchObjectField",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          },
          {
            "in": "path",
            "name": "field",
            "type": "string",
            "description": "The name of the field to set, see profile for a list of field names",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "The value being set violated the profile.",
            "schema": {
              "$ref": "#/definitions/ValidationErrorDocument"
            }
          },
          "404": {
            "description": "Research object not found"
          }
        }
      },
      "post": {
        "tags": [
          "research_object"
        ],
        "summary": "Append to Research Object field",
        "description": "Append a value to a list field on the Research Object",
        "operationId": "appendResearchObjectField",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          },
          {
            "in": "path",
            "name": "field",
            "type": "string",
            "description": "The name of the field to set, see profile for a list of field names. Must be a list field.",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "The value being appended violated the profile.",
            "schema": {
              "$ref": "#/definitions/ValidationErrorDocument"
            }
          },
          "404": {
            "description": "Research object not found"
          }
        }
      },
      "delete": {
        "tags": [
          "research_object"
        ],
        "summary": "Clear Research Object field",
        "description": "Reset a single field on the Research Object to its default value ({}, [] or null).",
        "operationId": "clearResearchObjectField",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          },
          {
            "in": "path",
            "name": "field",
            "type": "string",
            "description": "The name of the field to clear, see profile for a list of field names",
            "required": true
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "Research object not found"
          }
        }
      }
    },
    "/research_objects/{id}/content": {
      "patch": {
        "tags": [
          "research_object"
        ],
        "summary": "Patch Research Object content",
        "description": "Modify the Research Object's content using a JSON Patch",
        "operationId": "patchResearchObjectContent",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The JSON Patch to perform",
            "schema": {
              "$ref": "#/definitions/JSONPatchRequest"
            }
          },
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          }
        ],
        "consumes": [
          "application/json-patch+json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "The patched Research Object content violated the profile.",
            "schema": {
              "$ref": "#/definitions/ValidationErrorDocument"
            }
          },
          "404": {
            "description": "Research object not found"
          }
        }
      },
      "get": {
        "tags": [
          "research_object"
        ],
        "summary": "Get Research Object content",
        "description": "Get the Research Object's content",
        "operationId": "getResearchObjectContent",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "The numeric ID of the Research Object",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Research object not found"
          }
        }
      }
    }
  },
  "definitions": {
    "ProfileList": {
      "allOf": [
        {
          "$ref": "#/definitions/HALCollection"
        },
        {
          "type": "object",
          "properties": {
            "_embedded": {
              "type": "object",
              "properties": {
                "researchObjectProfileList": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Profile"
                  }
                }
              }
            }
          }
        }
      ]
    },
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "fields": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "_links": {
          "allOf": [
            {
              "$ref": "#/definitions/Links"
            },
            {
              "type": "object",
              "properties": {
                "researchObjects": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          ]
        }
      }
    },
    "ResearchObjectSummaryList": {
      "allOf": [
        {
          "$ref": "#/definitions/HALCollection"
        },
        {
          "type": "object",
          "properties": {
            "_embedded": {
              "type": "object",
              "properties": {
                "researchObjectSummaryList": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ResearchObjectSummary"
                  }
                }
              }
            }
          }
        }
      ]
    },
    "ResearchObject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "profileName": {
          "type": "string"
        },
        "content": {
          "type": "object"
        },
        "_links": {
          "allOf": [
            {
              "$ref": "#/definitions/Links"
            },
            {
              "type": "object",
              "properties": {
                "profile": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          ]
        }
      }
    },
    "ResearchObjectSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "profileName": {
          "type": "string"
        },
        "_links": {
          "$ref": "#/definitions/Links"
        }
      }
    },
    "HALCollection": {
      "type": "object",
      "properties": {
        "_embedded": {
          "type": "object"
        },
        "_links": {
          "$ref": "#/definitions/Links"
        }
      }
    },
    "Links": {
      "type": "object",
      "properties": {
        "self": {
          "type": "string",
          "format": "uri"
        }
      }
    },
    "JSONPointer": {
      "type": "string"
    },
    "JSONPatchRequest": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/JSONPatchDocument"
      }
    },
    "JSONPatchDocument": {
      "description": "A JSON Patch document (RFC 6902)",
      "required": [
        "op",
        "path"
      ],
      "properties": {
        "op": {
          "type": "string",
          "description": "The operation to be performed",
          "enum": [
            "add",
            "remove",
            "replace",
            "move",
            "copy",
            "test"
          ]
        },
        "path": {
          "allOf": [
            {
              "$ref": "#/definitions/JSONPointer"
            },
            {
              "description": "A pointer to the value being operated on. In the case of `move` and `copy` operations, it is a pointer to the destination."
            }
          ]
        },
        "value": {
          "description": "The value to be used within the operations. Can be any JSON type."
        },
        "from": {
          "allOf": [
            {
              "$ref": "#/definitions/JSONPointer"
            },
            {
              "description": "In the case of `move` and `copy` operations, a pointer to the source value."
            }
          ]
        }
      }
    },
    "ValidationErrorDocument": {
      "description": "A document detailing errors that occurred when trying to validate the Research Object, or a field on that Research Object.",
      "required": [
        "schemaLocation",
        "pointerToViolation",
        "causingExceptions"
      ],
      "properties": {
        "schemaLocation": {
          "allOf": [
            {
              "$ref": "#/definitions/JSONPointer"
            },
            {
              "description": "A pointer to the schema that was violated by this error."
            }
          ]
        },
        "pointerToViolation": {
          "allOf": [
            {
              "$ref": "#/definitions/JSONPointer"
            },
            {
              "description": "A pointer to the value in the request that caused the error."
            }
          ]
        },
        "causingExceptions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationErrorDocument"
          },
          "description": "Nested exceptions"
        },
        "keyword": {
          "type": "string",
          "description": "The property of the schema that was violated."
        },
        "message": {
          "type": "string",
          "description": "A human-readable error message."
        }
      }
    }
  }
}